+++ docs/release_notes.rst
@@ -23,6 +23,7 @@ Release Notes
 
 **Of interest to the User**:
 
+- PR #997: Fix bug causing symmetric auto wedge finding to fail. Add mirrored wedge to auto wedge fit function plot.
 - PR #998: remove the TOF offset that is done by the data aquisition system
 - PR #994: Remove unused module `drtsans/tof/eqsans/reduce.py`
 - PR #993: Skip slices with too high transmission error when using time sliced sample transmission run